PubChemCrawler

PubChemCrawler makes it easier to search the PubChem database from Julia. You can use it to access information about particular compounds or query substructures.

The package supports only a subset of the available functionality, but it is fairly straightforward to expand to other types of query. Pull requests are welcome! If you do want to make improvements to this package, this tutorial might help you get started.

Before you start: be aware of search limits

PubChem places significant limits on requests:

No more than 5 requests per second

No more than 400 requests per minute

No longer than 300 second running time per minute

Requests made via REST (the API targeted by this package) time out after 30s. There seems to be some caching, though, so sometimes it works to just resubmit the request.

The latter constraint is worth a bit of thought. Substructure search requests can sometimes generate hundreds of thousands of hits, and asking for lots of different properties for many different compounds can prolong the request beyond the timeout limit. One recommendation is to pursue a "staged" approach, where you first request just a few simple properties (e.g., MolecularFormula rather than the full SMILES string) and use that information to narrow the list of candidates. With a smaller list, it becomes feasible to request more detailed information.

Getting started

Some queries make use of the CID, the Compound IDentifier, which you can obtain in a variety of ways. Let's get the CID for aspirin:

julia> cid = get_cid(name="aspirin")
2244

julia> cid = get_cid(smiles="CC(=O)OC1=CC=CC=C1C(=O)O")   # use the SMILES string
2244

You can then retrieve a list of other properties:

julia> df = CSV.File(get_for_cids(cid; properties="MolecularFormula,MolecularWeight,XLogP,IsomericSMILES", output="CSV")) |> DataFrame
1×5 DataFrame
│ Row │ CID   │ MolecularFormula │ MolecularWeight │ XLogP   │ IsomericSMILES           │
│     │ Int64 │ String           │ Float64         │ Float64 │ String                   │
├─────┼───────┼──────────────────┼─────────────────┼─────────┼──────────────────────────┤
│ 1   │ 2244  │ C9H8O4           │ 180.16          │ 1.2     │ CC(=O)OC1=CC=CC=C1C(=O)O │

You can query such properties for a list of cids.

API

Queries

PubChemCrawler.get_cid — Function

cid = get_cid(name="glucose")
cid = get_cid(smiles="glucose")

Return the PubChem compound identification number for the specified compound.

PubChemCrawler.query_substructure — Function

msg = query_substructure(;cid=nothing, smiles=nothing,                           # specifier for the substructure to search for
                          properties="MolecularFormula,MolecularWeight,XLogP,",  # properties to retrieve
                          output="CSV")                                          # output format

Perform a substructure search of the entire database. You can specify the target via its cid, the SMILES string, or its name. Specify the properties you want to retrieve as a comma-separated list from among the choices in http://pubchemdocs.ncbi.nlm.nih.gov/pug-rest, "Compound Property Tables".

The output is a Vector{UInt8}. For output="CSV", a good choice to generate a manipulable result is DataFrame(CSV.File(msg)) from the DataFrames and CSV packages, respectively. Alternatively String(msg) will convert it to a string, which you can write to a file.

Example

julia> using PubChemCrawler, CSV, DataFrames

julia> cid = get_cid(name="estriol")
5756

julia> df = CSV.File(query_substructure(;cid)) |> DataFrame      # on Julia 1.0, use `(;cid=cid)`
11607×4 DataFrame
│ Row  │ CID       │ MolecularFormula │ MolecularWeight │ XLogP    │
│      │ Int64     │ String           │ Float64         │ Float64? │
├──────┼───────────┼──────────────────┼─────────────────┼──────────┤
│ 1    │ 5756      │ C18H24O3         │ 288.4           │ 2.5      │
│ 2    │ 5281904   │ C24H32O9         │ 464.5           │ 1.1      │
│ 3    │ 27125     │ C18H24O4         │ 304.4           │ 1.5      │
...

will query for derivatives of estriol.

PubChemCrawler.get_for_cids — Function

msg = get_for_cids(cids; properties=nothing, xrefs=nothing, output="CSV")

Retrieve the given properties or xrefs for a list of compounds specified by their cids.

See query_substructure for information about the arguments and return value. The supported values for xrefs are available at https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest under "XRefs".

Example

julia> using PubChem, JSON3

julia> cids = [get_cid(name="cyclic guanosine monophosphate"), get_cid(name="aspirin")]
2-element Array{Int64,1}:
 135398570
      2244

julia> dct = JSON3.read(get_for_cids(cids; xrefs="RN,", output="JSON"))   # get the Registry Number(s) (CAS)
JSON3.Object{Array{UInt8,1},Array{UInt64,1}} with 1 entry:
  :InformationList => {…

julia> dct[:InformationList][:Information]
2-element JSON3.Array{JSON3.Object,Array{UInt8,1},SubArray{UInt64,1,Array{UInt64,1},Tuple{UnitRange{Int64}},true}}:
 {
   "CID": 135398570,
    "RN": [
            "40732-48-7",
            "7665-99-8"
          ]
}
 {
   "CID": 2244,
    "RN": [
            "11126-35-5",
            "156865-15-5",
            "50-78-2",
            "52080-78-1",
            "921943-73-9",
            "98201-60-6",
            "99512-66-0"
          ]
}

Utilities

PubChemCrawler.parse_formula — Function

atomcounts = parse_formula(str::AbstractString)

Parse str as a chemical formula, return a list of atom=>multiplicity pairs.

Example

julia> parse_formula("C2CaH2O6")
4-element Array{Pair{String,Int64},1}:
  "C" => 2
 "Ca" => 1
  "H" => 2
  "O" => 6

PubChemCrawler.atomregex — Function

rex = atomregex(chemicalsymbol)

Create a regular expression for detecting how many atoms of type chemicalsymbol are in a molecular formula.

Examples

The formula for calcium bicarbonate is Ca(HCO3)2, i.e., C2CaH2O6.

julia> match(atomregex("C"), "C2CaH2O6")
RegexMatch("C2", 1="2")

julia> match(atomregex("Ca"), "C2CaH2O6")
RegexMatch("Ca", 1="")

julia> match(atomregex("H"), "C2CaH2O6")
RegexMatch("H2", 1="2")

julia> match(atomregex("O"), "C2CaH2O6")
RegexMatch("O6", 1="6")

Note that the regex for "C" does not match "Ca".